+++ scalapb-runtime/js/src/main/scala/com/google/protobuf/descriptor/EnumValueOptions.scala
@@ -18,8 +18,9 @@ package com.google.protobuf.descriptor
 @SerialVersionUID(0L)
 final case class EnumValueOptions(
     deprecated: scala.Option[Boolean] = None,
-    uninterpretedOption: _root_.scala.collection.Seq[com.google.protobuf.descriptor.UninterpretedOption] = _root_.scala.collection.Seq.empty
-    ) extends com.trueaccord.scalapb.GeneratedMessage with com.trueaccord.scalapb.Message[EnumValueOptions] with com.trueaccord.lenses.Updatable[EnumValueOptions] {
+    uninterpretedOption: _root_.scala.collection.Seq[com.google.protobuf.descriptor.UninterpretedOption] = _root_.scala.collection.Seq.empty,
+    unknownFields: _root_.scalapb.UnknownFieldSet = _root_.scalapb.UnknownFieldSet()
+    ) extends com.trueaccord.scalapb.GeneratedMessage with com.trueaccord.scalapb.Message[EnumValueOptions] with com.trueaccord.lenses.Updatable[EnumValueOptions] with _root_.com.trueaccord.scalapb.ExtendableMessage[EnumValueOptions] {
     @transient
     private[this] var __serializedSizeCachedValue: Int = 0
     private[this] def __computeSerializedValue(): Int = {
@@ -49,6 +50,7 @@ final case class EnumValueOptions(
     def mergeFrom(`_input__`: _root_.com.google.protobuf.CodedInputStream): com.google.protobuf.descriptor.EnumValueOptions = {
       var __deprecated = this.deprecated
       val __uninterpretedOption = (_root_.scala.collection.immutable.Vector.newBuilder[com.google.protobuf.descriptor.UninterpretedOption] ++= this.uninterpretedOption)
+      val _unknownFields__ = new _root_.scalapb.UnknownFieldSet.Builder(this.unknownFields)
       var _done__ = false
       while (!_done__) {
         val _tag__ = _input__.readTag()
@@ -58,12 +60,13 @@ final case class EnumValueOptions(
             __deprecated = Some(_input__.readBool())
           case 7994 =>
             __uninterpretedOption += _root_.com.trueaccord.scalapb.LiteParser.readMessage(_input__, com.google.protobuf.descriptor.UninterpretedOption.defaultInstance)
-          case tag => _input__.skipField(tag)
+          case tag => _unknownFields__.parseField(tag, _input__)
         }
       }
       com.google.protobuf.descriptor.EnumValueOptions(
           deprecated = __deprecated,
-          uninterpretedOption = __uninterpretedOption.result()
+          uninterpretedOption = __uninterpretedOption.result(),
+          unknownFields = _unknownFields__.result()
       )
     }
     def getDeprecated: Boolean = deprecated.getOrElse(false)