+++ scalapb-runtime/js/src/main/scala/com/google/protobuf/descriptor/MessageOptions.scala
@@ -66,8 +66,9 @@ final case class MessageOptions(
     noStandardDescriptorAccessor: scala.Option[Boolean] = None,
     deprecated: scala.Option[Boolean] = None,
     mapEntry: scala.Option[Boolean] = None,
-    uninterpretedOption: _root_.scala.collection.Seq[com.google.protobuf.descriptor.UninterpretedOption] = _root_.scala.collection.Seq.empty
-    ) extends com.trueaccord.scalapb.GeneratedMessage with com.trueaccord.scalapb.Message[MessageOptions] with com.trueaccord.lenses.Updatable[MessageOptions] {
+    uninterpretedOption: _root_.scala.collection.Seq[com.google.protobuf.descriptor.UninterpretedOption] = _root_.scala.collection.Seq.empty,
+    unknownFields: _root_.scalapb.UnknownFieldSet = _root_.scalapb.UnknownFieldSet()
+    ) extends com.trueaccord.scalapb.GeneratedMessage with com.trueaccord.scalapb.Message[MessageOptions] with com.trueaccord.lenses.Updatable[MessageOptions] with _root_.com.trueaccord.scalapb.ExtendableMessage[MessageOptions] {
     @transient
     private[this] var __serializedSizeCachedValue: Int = 0
     private[this] def __computeSerializedValue(): Int = {
@@ -112,6 +113,7 @@ final case class MessageOptions(
       var __deprecated = this.deprecated
       var __mapEntry = this.mapEntry
       val __uninterpretedOption = (_root_.scala.collection.immutable.Vector.newBuilder[com.google.protobuf.descriptor.UninterpretedOption] ++= this.uninterpretedOption)
+      val _unknownFields__ = new _root_.scalapb.UnknownFieldSet.Builder(this.unknownFields)
       var _done__ = false
       while (!_done__) {
         val _tag__ = _input__.readTag()
@@ -127,7 +129,7 @@ final case class MessageOptions(
             __mapEntry = Some(_input__.readBool())
           case 7994 =>
             __uninterpretedOption += _root_.com.trueaccord.scalapb.LiteParser.readMessage(_input__, com.google.protobuf.descriptor.UninterpretedOption.defaultInstance)
-          case tag => _input__.skipField(tag)
+          case tag => _unknownFields__.parseField(tag, _input__)
         }
       }
       com.google.protobuf.descriptor.MessageOptions(
@@ -135,7 +137,8 @@ final case class MessageOptions(
           noStandardDescriptorAccessor = __noStandardDescriptorAccessor,
           deprecated = __deprecated,
           mapEntry = __mapEntry,
-          uninterpretedOption = __uninterpretedOption.result()
+          uninterpretedOption = __uninterpretedOption.result(),
+          unknownFields = _unknownFields__.result()
       )
     }
     def getMessageSetWireFormat: Boolean = messageSetWireFormat.getOrElse(false)