+++ scalapb-runtime/jvm/src/main/scala/com/google/protobuf/descriptor/FileOptions.scala
@@ -94,8 +94,9 @@ final case class FileOptions(
     ccEnableArenas: scala.Option[Boolean] = None,
     objcClassPrefix: scala.Option[String] = None,
     csharpNamespace: scala.Option[String] = None,
-    uninterpretedOption: _root_.scala.collection.Seq[com.google.protobuf.descriptor.UninterpretedOption] = _root_.scala.collection.Seq.empty
-    ) extends com.trueaccord.scalapb.GeneratedMessage with com.trueaccord.scalapb.Message[FileOptions] with com.trueaccord.lenses.Updatable[FileOptions] {
+    uninterpretedOption: _root_.scala.collection.Seq[com.google.protobuf.descriptor.UninterpretedOption] = _root_.scala.collection.Seq.empty,
+    unknownFields: _root_.scalapb.UnknownFieldSet = _root_.scalapb.UnknownFieldSet()
+    ) extends com.trueaccord.scalapb.GeneratedMessage with com.trueaccord.scalapb.Message[FileOptions] with com.trueaccord.lenses.Updatable[FileOptions] with _root_.com.trueaccord.scalapb.ExtendableMessage[FileOptions] {
     @transient
     private[this] var __serializedSizeCachedValue: Int = 0
     private[this] def __computeSerializedValue(): Int = {
@@ -190,6 +191,7 @@ final case class FileOptions(
       var __objcClassPrefix = this.objcClassPrefix
       var __csharpNamespace = this.csharpNamespace
       val __uninterpretedOption = (_root_.scala.collection.immutable.Vector.newBuilder[com.google.protobuf.descriptor.UninterpretedOption] ++= this.uninterpretedOption)
+      val _unknownFields__ = new _root_.scalapb.UnknownFieldSet.Builder(this.unknownFields)
       var _done__ = false
       while (!_done__) {
         val _tag__ = _input__.readTag()
@@ -225,7 +227,7 @@ final case class FileOptions(
             __csharpNamespace = Some(_input__.readString())
           case 7994 =>
             __uninterpretedOption += _root_.com.trueaccord.scalapb.LiteParser.readMessage(_input__, com.google.protobuf.descriptor.UninterpretedOption.defaultInstance)
-          case tag => _input__.skipField(tag)
+          case tag => _unknownFields__.parseField(tag, _input__)
         }
       }
       com.google.protobuf.descriptor.FileOptions(
@@ -243,7 +245,8 @@ final case class FileOptions(
           ccEnableArenas = __ccEnableArenas,
           objcClassPrefix = __objcClassPrefix,
           csharpNamespace = __csharpNamespace,
-          uninterpretedOption = __uninterpretedOption.result()
+          uninterpretedOption = __uninterpretedOption.result(),
+          unknownFields = _unknownFields__.result()
       )
     }
     def getJavaPackage: String = javaPackage.getOrElse("")