+++ scalapb-runtime/shared/src/main/scala/com/trueaccord/scalapb/GeneratedExtension.scala
@@ -1,55 +1,66 @@
 package com.trueaccord.scalapb
 
-import com.google.protobuf.GeneratedMessageV3.ExtendableMessage
-import com.google.protobuf.{ByteString, UnknownFieldSet}
-import com.trueaccord.scalapb.GeneratedExtension.JavaExtendee
+import com.google.protobuf.{ByteString, CodedInputStream, InvalidProtocolBufferException}
 
-case class GeneratedExtension[J <: JavaExtendee[J], +T](get: J => T)
+import _root_.scalapb.UnknownFieldSet
+
+case class GeneratedExtension[C <: ExtendableMessage[C], +T](get: C => T)
 
 object GeneratedExtension {
-  type JavaExtendee[S <: ExtendableMessage[S]] = com.google.protobuf.Message with com.google.protobuf.GeneratedMessageV3.ExtendableMessage[S]
-
-  import scala.collection.JavaConverters._
-
   /* To be used only be generated code */
   def readMessageFromByteString[T <: GeneratedMessage with Message[T]](cmp: GeneratedMessageCompanion[T])(bs: ByteString): T = {
     cmp.parseFrom(bs.newCodedInput())
   }
 
-  def forSingularUnknownField[J <: JavaExtendee[J], E, T](
-    fieldNumber: Int, listGetter: UnknownFieldSet.Field => java.util.List[E])(convert: E => T, default: T): GeneratedExtension[J, T] =
+  def forSingularUnknownField[C <: ExtendableMessage[C], E, T](
+    fieldNumber: Int, listGetter: UnknownFieldSet.Field => Vector[E])(convert: E => T, default: T): GeneratedExtension[C, T] =
     GeneratedExtension {
       extendee =>
-        val list = listGetter(extendee.getUnknownFields.getField(fieldNumber))
-        if (list.isEmpty) default
-        else convert(list.get(0))
+        (for {
+          field <- extendee.unknownFields.getField(fieldNumber)
+          elem <- listGetter(field).lastOption
+        } yield convert(elem)).getOrElse(default)
     }
 
-  def forOptionalUnknownField[J <: JavaExtendee[J], E, T](
-    fieldNumber: Int, listGetter: UnknownFieldSet.Field => java.util.List[E])(convert: E => T): GeneratedExtension[J, Option[T]] =
+  def forOptionalUnknownField[C <: ExtendableMessage[C], E, T](
+    fieldNumber: Int, listGetter: UnknownFieldSet.Field => Vector[E])(convert: E => T): GeneratedExtension[C, Option[T]] =
     GeneratedExtension {
       extendee =>
-        val list = listGetter(extendee.getUnknownFields.getField(fieldNumber))
-        if (list.isEmpty) None
-        else Some(convert(list.get(0)))
+        extendee.unknownFields.getField(fieldNumber) match {
+          case None => None
+          case Some(field) => listGetter(field).lastOption.map(convert)
+        }
     }
 
 
-  def forRepeatedUnknownField[J <: JavaExtendee[J], E, T](
-    fieldNumber: Int, listGetter: UnknownFieldSet.Field => java.util.List[E])(convert: E => T): GeneratedExtension[J, Vector[T]] = {
+  def forRepeatedUnknownField[C <: ExtendableMessage[C], E, T](
+    fieldNumber: Int, listGetter: UnknownFieldSet.Field => Vector[E], unpack: CodedInputStream => E)(convert: E => T): GeneratedExtension[C, Vector[T]] = {
     GeneratedExtension {
-      extendee: J =>
-        listGetter(extendee.getUnknownFields.getField(fieldNumber)).asScala.map(convert)(scala.collection.breakOut)
+      extendee: C =>
+        extendee.unknownFields.getField(fieldNumber) match {
+          case None => Vector.empty
+          case Some(field) =>
+            extendee.unknownFields.getField(fieldNumber) match {
+              case None => Vector.empty
+              case Some(field) =>
+                if (field.lengthDelimited.nonEmpty && unpack != null) {
+                  if (listGetter(field).nonEmpty) {
+                    throw new InvalidProtocolBufferException("Mixed packed and unpacked data.")
+                  }
+                  val v = Vector.newBuilder[T]
+                  field.lengthDelimited.foreach {
+                    ld =>
+                      val ci = CodedInputStream.newInstance(ld.toByteArray)
+                      while (ci.getBytesUntilLimit > 0) {
+                        v += convert(unpack(ci))
+                      }
+                  }
+                  v.result()
+                } else {
+                  listGetter(field).map(convert)
+                }
+            }
+        }
     }
-  }
-
-  def forExtension[J <: JavaExtendee[J], T](convert: J => T) = {
-    GeneratedExtension(convert)
-  }
-}
-
-object Implicits {
-  implicit class JavaMessageExt[J <: JavaExtendee[J]](val javaExtendee: J) extends AnyVal {
-    def extension[T](ext: GeneratedExtension[J, T]): T = ext.get(javaExtendee)
   }
 }